build: Fix GitHub download URLs to pin versions and reduce need for override
The file name at the end of a github.com/<user>/<repo>/archive URL is
arbitrary, so we shouldn't need the 'override' feature of fetch.py for
GitHub downloads - we can just ask for the filename we want.   However
fetching from github.com/<user>/<repo>/archive seems to return the most
recent version, not necessarily the version we want (the version in the
filename is ignored, because GitHub doesn't care about the filename).
Therefore we should request archive/<tag>/filename-1.2.3.tar.gz,
to make sure we always get the same version.

Signed-off-by: Euan Harris <euan.harris@citrix.com>