build: Add a naive mechanism for building packages from local sources
To include locally-modified source in a package:

  1. Check out the source:
       git clone https://github.com/euanh/xenserver-install-wizard.git

  2. Edit the spec file to refer to this source tree.   There is a
     specific format which must be followed:
       file:///path/to/checkout#/tarball-version.tar.gz
     Note the slash after the # - this is vital to make rpmbuild ignore
     the rest of the URL.

     In this case:
       Source0: file:///scratch/xenserver-install-wizard#/%{name}-%{version}

  3. Re-run makemake.py and build
       ./makemake.py > Makefile
       make xenserver-install-wizard

The RPM will use the version number from the spec file.   If you have already
built and installed a package with this version, you will need to force RPM
to install the new one.   Alternatively, you can increment the minor version
number when you edit the spec file to point at the local tree.

Signed-off-by: Euan Harris <euan.harris@citrix.com>