build: Ensure all spec files have mandatory %files list
A spec file must have a %files list, even if the list is empty.  Some of
our base packages don't have %files lists, which causes mock to build
them and throw them away.   We still generate Makefile rules to build
these files, so they are rebuilt and thrown away every time we run make.

For library packages, we should move the libraries into the base package;
keep only the headers, interfaces or other developer-focussed files
in the -devel package, and have the -devel package depend on the base
package.

Signed-off-by: Euan Harris <euan.harris@citrix.com>