forkexecd: Copy extra sources into build directory
All sources are copied into the SRPM and unpacked in $RPM_SOURCE_DIR.
Previously this spec file installed the forkexecd-init file directly
from $RPM_SOURCE_DIR into the output directory, but in our Debian build
flow we currently build the source package by executing the %prep phase
and then packaging up the build directory.   For this to work, we need to
make sure that we copy all the source files into the build directory in
the %prep section.

Signed-off-by: Euan Harris <euan.harris@citrix.com>