xsc-latest-snapshot: Create build dir and copy in sources
The Debian packaging script expects the %prep phase of
rpmbuild to create a build directory and copy all sources
into it.   Normally the directory is created by the setup
macro unpacking the source tarball, but since
xenserver-core-latest-snapshot doesn't have a source tarball
we must pass different options to have it create an empty
directory.

Signed-off-by: Euan Harris <euan.harris@citrix.com>