makedeb: Map ocaml-camomile-data to libcamomile-data
dh_ocaml insists that a package with 'ocaml' in its name
should have a -dev package.  Currently we call dh_ocaml
for all packages (we should only call it for packages that
actually contain OCaml code), so the lack of a
libcamomile-ocaml-data-dev file causes the build to fail.
As a workaround, map the package to libcamomile-data.

Signed-off-by: Euan Harris <euan.harris@citrix.com>