linux-guest-loader: Add build requirement on xcp-python-libs
linux-guest-loader does not require xcp-python-libs to be installed in
order to build it, but it does depend on it at run time.   xcp-python-libs
will still be built because it is a dependency of the 'all 'target, but
some other packages buildrequire linux-guest-loader, and since there is
no build dependency between linux-guest-loader and xcp-python-libs,
xcp-python-libs might not have been built at the time linux-guest-loader
is installed, so the installation will fail and stop the build.

Signed-off-by: Euan Harris <euan.harris@citrix.com>