Makefile: Enable concurrent RPM builds
Two problems prevent us from building RPMs concurrently:

 * By default, mock can only run one instance of a given
   chroot environment at a time.   This is easily fixed
   by using the --uniqueext option to name instances
   uniquely.

 * Createrepo needs to update the repository database
   files in SRPMS and RPMS after each package is built.
   It's not clear whether it's safe to run two instances
   of createrepo against the same database concurrently.
   To prevent this, we wrap the call to createrepo with
   flock.

Copying build products back to the RPMS and SRPMS directories
concurrently should be safe because each job builds a different
package with a different filename.  An exception to this is
the build log file, which is common to all jobs.   We should
switch to separate logging for each package.

Signed-off-by: Euan Harris <euan.harris@citrix.com>