Debian: Move repository metadata generation into Makefile
Debian package-building documents suggest that the chroot (pbuilder
or cowbuilder) should contain a startup script which regenerates the
package repository metadata.   In our case, the D05deps script runs
apt-ftparchive on the source and binary package directories.

This approach has the advantage that the repository metadata is guaranteed
to be up to date before the build starts.   In contrast, when building
RPMs we have to remember to call createrepo after building each package.
However the disadvantage is that regenerating the package metadata is
a side effect of building the package.   This will make it more difficult
to add a cache around Debian package building.

For this reason, we will move the metadata regeneration step out of the
chroot startup sequence and into the Makefile, mirroring the way we
build RPMs.

Signed-off-by: Euan Harris <euan.harris@citrix.com>