downloader: Update access time for existing archives
If a spec file is updated, make will call the downloader script to get
the source archives mentioned in the spec file.   If the source file
name has not changed and the file already exists, we need to update
its access time otherwise the spec file will always be newer than the
archives.   Otherwise make will always consider the source archives
to be out of date and will run the downloader to refresh them.

Signed-off-by: Euan Harris <euan.harris@citrix.com>